980 change point data chart
